backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-core
langsmith
openai
chromadb
faiss-cpu
pypdf
requests
tiktoken
python-dotenv
beautifulsoup4
lxml
tenacity
pydantic>=2
